backend/pyproject.toml
[project]
name = "labnote-backend"
version = "0.1.0"
description = "LabNote backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "pydantic-settings>=2.3.4",
  "python-multipart>=0.0.9",
  "pypdf>=4.2.0",
  "Pillow>=10.3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
